Added new features and improved code readability
A ButtonColor feature was added to allow button color change dynamically. The BoardContent class now has a method that gives the list of words. A new enumeration, State, was introduced for tracking various states of the game. The WSModel class has been considerably improved with more meaningful methods and better code readability. Also, added a ColorStack enum for representing color in RGB format. Fixed a test case in WSModelTest.